srv6int/requirements.txt
scapy-2.6.1